DOMUI Prefs: Replace search engine edit overlay with inline editing.
Validation feedback still needs polish, but this is completely functional, so I'll leave that for a follow-up.

This introduces a new shared class for editable lists; we should migrate the existing editable lists to it, expanding as necessary, but I didn't want to make this patch too unwieldly.

BUG=63825,61742
TEST=Search engines should be editable and addable inline in DOMUI prefs.

Review URL: http://codereview.chromium.org/6151004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71121 0039d316-1c4b-4281-b951-d872f2087c98