Improved tabbed options subsheets.
1. Fix subsheet RTL.

for the record, and to whom it may concern:

"""
If neither 'left' nor 'right' is 'auto', the position is over-constrained, and one of them has to be ignored. If the 'direction' property of the containing block is 'ltr, the value of 'left' wins and 'right' becomes -'left'. If 'direction' of the containing block is 'rtl', 'right' wins and 'left' is ignored.
"""

2. add handlers to the body element to remove the subsheets when the user clicks outside them.

BUG=67849
TEST=
1. LANGUAGE=he_IL.utf-8 ./out/Release/chrome (looks right)
2. clicking away from the subsheets to dismiss them should still work.
3. clicking to the right (in ltr) of the subsheets should dismiss them (this is new)

Review URL: http://codereview.chromium.org/6277005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@71724 0039d316-1c4b-4281-b951-d872f2087c98