Partial support for Mac plugin layering, using cutout rects
This is a temporary measure until we get real GPU compositing of accelerated plugins on the Mac; this adds support for the cutout-rect approach currently used by Windows and Linux to allow iframes to appear to render above windowed plugins. This means most sites that layer windowed plugins will now work on the Mac if they work on Windows and Linux.

BUG=38967
TEST=Play a video on YouTube on the Mac, then use the search field. The suggestion menu should draw over the video (however, see bug 71732 for a known issue).

Review URL: http://codereview.chromium.org/6349062

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@73633 0039d316-1c4b-4281-b951-d872f2087c98