Update GYP to r890 to pick up:
Tolerate an empty set of configurations by falling back to generate a Default
configuration. input.py handles this by allays offering a Default
configuration even if the .gyp file didn't specify any, but in a .gyp file
that offers no targets, there's no way for this to make it to the project
scope. xcode.py normally builds a configuration list at the project scope
based on all of the configuration names in all of the targets in an
.xcodeproj. Adding this check protects against that list being empty because
no targets are present.
Review URL: http://codereview.chromium.org/6516006

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/6520013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@74824 0039d316-1c4b-4281-b951-d872f2087c98