Added TranslateVariablesInPolicy to the policy toolkit.
This function was implemented for mac and win in the chrome_main_xxx files because it was used only by the UserDataDir policy until now. However we need this in the DownloadDirectory policy too so I extracted in in a more
global way in the policy::path_parser namespace.

BUG=None
TEST=unit_tests:PolicyPathParser.*

Review URL: http://codereview.chromium.org/6469081

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76045 0039d316-1c4b-4281-b951-d872f2087c98