Hover/selection of username dropdown items should do nothing
This changes the password autofill interaction between mouse-hover and menu item selection's interaction with the contents of the <input> text.  This restores the original behavior of only modifying the <input> text when the user has explicity "accepted" the selection in the menu, as opposed to hovering over the selection or key-down/up'ing to the selection.

BUG=74868
TEST=PasswordAutoFillManagerTest.SuggestionAccept

Review URL: http://codereview.chromium.org/6623005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76999 0039d316-1c4b-4281-b951-d872f2087c98