Refactor GetCommandLineArgumentsCommon to not use strings.
Passing around command lines as strings is dangerous, because a given
command line is interpretable only with respect to a given parser.
This changes a function to instead pass a CommandLine object for
carrying the arguments around, and relies on the callers to stringify
that as necessary.

It turns out that the Linux desktop file format has a well-defined
syntax that is different than the syntax we were using.

BUG=69467

Review URL: http://codereview.chromium.org/6624072

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77338 0039d316-1c4b-4281-b951-d872f2087c98