Normalize the top-level ui/ module.
This CL creates four .gypi files that are included in /ui/ui.gyp. This allows
for a single unittest binary for all the submodules, rather than having
individual unittest binaries for each submodule.

To not break the buildbot, this creates a fake target that copies ui_unittests
to gfx_unittests as temporary scaffolding.

BUG=72317
TEST=gfx_unittests (really ui_unittests) passes

Review URL: http://codereview.chromium.org/6688007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@78711 0039d316-1c4b-4281-b951-d872f2087c98