Implement CrashUploadListWin to enable chrome://crashes on Windows.
Added new class CrashUploadListWin, subclass of CrashUploadList which
has the Windows platform-specific logic to read crash logs from Windows
Event Log.

Added static CrashUploadList::Create() factory method which returns an
instance of CrashUploadListWin on Windows and an instance of
CrashUploadList on other platforms.

Also implemented the max_count parameter in
CrashUploadList::GetUploadedCrashes(), which was part of
the interface but was ignored by the code.

BUG=73468
TEST=Open chrome://crashes under Windows after one or more
     crash reports were uploaded. See the crash reports listed.


Review URL: http://codereview.chromium.org/6771021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@80165 0039d316-1c4b-4281-b951-d872f2087c98