Rework the in-browser thumbnailing by doing everything in ThumbnailGenerator.
Before this change, the thumbnail generation occurred when a tab received
page contents from from a renderer (in TabContents::OnPageContents), but
this is not the right place to do so (ex. the page may not be yet fully
loaded, the tab may be hidden, etc.)

After this change, the thumbnail generation will be done based on heuristics
implemented in ThumbnailGenerator. The current heuristics is as simple as
attempting to update thumbnails when a renderer is about to be hidden, but
it's just a starting point. We can tweak heuristics to get better quality
thumbnails.

Along the way, remove the complicated logic in ThumbnailGenerator that was
used for the overview mode on Chrome OS. We no longer use have that feature.

BUG=65936
TEST=run chrome with --enable-in-browser-thumbnailing and confirm that thumbnails are updated as intended.

Review URL: http://codereview.chromium.org/6810044

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@81224 0039d316-1c4b-4281-b951-d872f2087c98