Place focus in password field when joining Wi-Fi network
Be smarter about keyboard focus when joining a Wi-Fi network, specifically, put focus in the SSID field if that is required, or the EAP access method if that is required, or (most typically) the password field.

Note that this does not put focus in the password field when the user types the password incorrectly and the dialog re-spawns.  I haven't figured out how to do that yet.

BUG=chromium-os:13767
TEST=Join a password protected Wi-Fi network.  When the password dialog appears, verify that focus is in the password text field.

Review URL: http://codereview.chromium.org/6858007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@81817 0039d316-1c4b-4281-b951-d872f2087c98