[Chrome OS] Reimplement SignedSettings in terms of device policies.
Owner settings are being re-implemented as a single, signed protobuf
that is stored by the chromium os session_manager.  Thus, to write a
setting, you need to have the existing policy, update it, re-sign it,
and then have it stored.  This could be done by requesting the policy
every time, or by caching it and updating it upon every successful
store.  We cache it in the OwnershipService singleton.  The SignedSettings
API remains unchanged, semantically.

This should all be re-implemented more cleanly (remove the out-dated notions
of storing/retrieving properties and whitelisting, etc) as part of
http://crosbug.com/14054

BUG=chromium-os:13230
TEST=unit tests

Review URL: http://codereview.chromium.org/6831011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@81828 0039d316-1c4b-4281-b951-d872f2087c98