[Sync] Refactor ExtensionService for sync
Move sync-related methods in ExtensionService to the interface
ExtensionServiceSyncInterface.  Add methods GetSyncData() and
GetSyncDataList() for sync to query.

Make extension/app sync code use new interface and methods.  Stop
listening to EXTENSION_INSTALLED notification as it is unneeded;
EXTENSION_LOADED/EXTENSION_UPDATE_DISABLED is called afterwards.

Clean up now-unneeded methods in ExtensionServiceInterface.

Stop keeping track of unsynced extensions in SlurpClientData().  We check it in SlurpServerData(), but we shouldn't be sending that data to the server in the first place.

Add OVERRIDE annotations to subclasses of ExtensionService* subclasses.

BUG=77995
TEST=

Review URL: http://codereview.chromium.org/6900012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@82755 0039d316-1c4b-4281-b951-d872f2087c98