Replace the virtual InfoBarDelegate::InfoBarClosed() function with a non-virtual one.  This is a step along the way to killing it entirely.  This also adds a lot of OVERRIDE markers and does some other cleanup in a few places.
The original (stupid) design for the delegate class left subclasses great flexibility in how they mapped infobars to delegates.  In practice, no one ever wanted multiple infobars driven off a single delegate, so the mapping was always one-to-one.  As a result, it was always correct for InfoBarClosed() to "delete this", but because the base class did not do so, every subclass needed to.  Most did; the others leaked memory and failed to run their destructors.

This change forces the base class to delete itself.  This fixes the delegate leaks in the couple subclasses that failed to do this.  It also eliminates a lot of copy-and-pasted "delete this" implementations.

Ultimately, we'll be moving to a model where the InfoBar "view" class owns the delegate and deletes it directly, which will eliminate InfoBarClosed() completely.

BUG=62154
TEST=none
Review URL: http://codereview.chromium.org/6926001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@84195 0039d316-1c4b-4281-b951-d872f2087c98